+++ superset/tasks/caching/strategy.py
@@ -1,0 +1,49 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+# pylint: disable=too-few-public-methods
+
+from typing import List
+
+
+class Strategy:
+    """
+    A cache warm up strategy.
+
+    Each strategy defines a `get_urls` method that returns a list of URLs to
+    be fetched from the `/superset/warm_up_cache/` endpoint.
+
+    Strategies can be configured in `superset/config.py`:
+
+        CELERYBEAT_SCHEDULE = {
+            'cache-warmup-hourly': {
+                'task': 'cache-warmup',
+                'schedule': crontab(minute=1, hour='*'),  # @hourly
+                'kwargs': {
+                    'strategy_name': 'top_n_dashboards',
+                    'top_n': 10,
+                    'since': '7 days ago',
+                },
+            },
+        }
+
+    """
+
+    def __init__(self) -> None:
+        pass
+
+    def get_urls(self) -> List[str]:
+        raise NotImplementedError("Subclasses must implement get_urls!")
